% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_count.R
\name{gen_count}
\alias{gen_count}
\title{gen_count}
\usage{
gen_count(text)
}
\arguments{
\item{text}{(string): A string object.}
}
\value{
data.frame with each word from the dictionary matched with the text and its number of occurances.
}
\description{
Counts the number of masculine and feminine words in the document
}
\details{
Takes the number of words that are loosely categorized as Masculine, Feminine, or Neutral based on Roberts and Utych's (2019) definition.
Feminine words had a score below 2.5, Neutral words had a score higher than 2.5 and lower than 5.5, Masculine words had a score higher than 5.5.
}
\examples{
text <- 'This person was a heroine due to their fighting during the war.'
result_df <- genCountR::gen_count(text)

}
