% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_score.R
\name{gen_score}
\alias{gen_score}
\title{gen_score}
\usage{
gen_score(text)
}
\arguments{
\item{text}{(string): A string object}
}
\value{
list object with avg_score of the supplied text string, total_score of supplied text string, and data.frame of matches
}
\description{
Calculates the score of the supplied text string based on the Gendered Language Dictionary created by Roberts and Utych (2019).
}
\details{
Takes the matched words and their occurrence in the supplied text, finds the score for those matched words in the dictionary, sums those scores up and then divides it by the total number of words in the dictionary.
}
\examples{
text <- 'Hero. hero Heroine. heroine, Prison. Prisom.'
result <- genCountR::gen_score(text)
result$avg_score
result$total_score
result$df

}
