% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genCountR-package.R
\docType{package}
\name{genCountR-package}
\alias{genCountR}
\alias{genCountR-package}
\title{genCountR: Interacting with Roberts and Utych's (2019) Gendered Language Dictionary}
\description{
Allows users to generate a gendered language score according to the gendered language dictionary in Roberts and Utych (2019) \doi{10.1177/1065912919874883}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gencounter.app.damoncroberts.com}
  \item \url{https://damoncharlesroberts.github.io/genCountR/}
}

}
\author{
\strong{Maintainer}: Damon Roberts \email{damon.charles.roberts@gmail.com} (\href{https://orcid.org/0000-0002-4360-3675}{ORCID})

}
\keyword{internal}
