% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_clean.R
\name{text_clean}
\alias{text_clean}
\title{text_clean}
\usage{
text_clean(text)
}
\arguments{
\item{text}{(string): A string object.}
}
\value{
list of each word in all lower case and without punctuation.
}
\description{
Cleans the supplied text string and converts it into a list of individual words.
}
\details{
Takes the string, converts all the characters to lower case, removes punctuation, and splits the string into individual words.
}
