% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_count.R
\name{word_count}
\alias{word_count}
\title{word_count}
\usage{
word_count(word_item)
}
\arguments{
\item{word_item}{(vector or array): A vector or array of words from the text wanting to be matched to dictionary.}
}
\value{
data.frame object of with count of masculine, feminine, and masculine words.
}
\description{
Count number of times a particular word from the dictionary shows up in a document.
}
\details{
Takes the number of words that are loosely categorized as Masculine, Feminine, or Neutral based on Roberts and Utych's (2019) definition.
Feminine words had a score below 2.5, Neutral words had a score higher than 2.5 and lower than 5.5, Masculine words had a score higher than 5.5.
}
